teams/asahi: FEX: Clean up some redundant FEX packages
ibus: This is an input method, which is a user choice. It doesn't make
sense to ship some input method libs and not others. These libs
typically communicate with the main IM daemon over UNIX sockets, so this
won't work anyway without additional work. (This might cause some
spurious errors as the guest tries to load the nonexistent IM, but we
can ignore them since it wouldn't work anyway).

xdg-desktop-portal: This is just data and executables, no libraries, so
it makes no sense for FEX.

glibc-langpack-en, kbd-misc: Only needed for kiwi, add a comment

zenity: Used by Steam but it's just a binary, so this should be an
aarch64 dependency, not run emulated.

llvm 14 & 15: I don't think there's a good reason to ship these? Any
reasonable binary-packaged app shouldn't be depending on specific LLVM
versions like this. Let's keep llvm18 & 19 since 18 is used by our Mesa
builds.

sudo & rsync: Binaries only (and sudo doesn't even work, neither under
emulation nor plain in muvm).

dbus: What we really need is the libs, so replace with that. dbus
probably won't do what we want inside the container, but for many use
cases dbus is not critical so it's okay for things to fail as long as
the libs are there.

spirv-tools-devel: Replace with spirv-tools-libs (dep)
glew-devel: Replace with libGLEW (dep)