grub2: Use `menu_auto_hide=1` to match anaconda
menu_auto_hide=1 is required for the boot_success provided by Fedora's
grub2 package. This will fall back to the previous kernel and show the
usually hidden grub menu when booting fails.

This makes it unnecessary to override the timeout_style and using a
reduced timeout of 1 second for disk images.