ELN: Adjust aarch64 images to use the correct serial console
The serial device name on Arm is ttyAMA, not ttyS. This splits the
configuration for ELN images based on the architecture and adjusts the
kernel command line.

Note that prior to this change, aarch64 would still output kernel logs
to the ttyAMA0 console (on Azure, at least) due to the ACPI SPCR table.
On aarch64 the console in this table, if it's present, is _always_
configured as the kernel's default console. However, systemd logs would
not be sent to the serial console because it seems it was honoring the
kernel command line configuration which had ttyS0 as the default.

Note: this is the exact same change as 4979662a547e ("Cloud: Adjust
aarch64 images to use the correct serial console"), if you're reverting
this, you probably want to also revert that one.

Also see: https://pagure.io/cloud-sig/issue/394
Also see: https://bugzilla.redhat.com/show_bug.cgi?id=1661288

Signed-off-by: Jeremy Cline <jeremycline@linux.microsoft.com>