teams/cloud/container: Don't explicitly install p11-kit for Toolbx
The p11-kit RPM is neither explicitly mentioned in the list of default
RPMs on Fedora Silverblue and Workstation [1], nor was it explicitly
installed by the Container/Dockerfile equivalent of the fedora-toolbox
OCI images [2].

The Container/Dockerfile listed p11-kit as one of the RPMs that had to
be reinstalled, because it was inherited from the fedora base image, and
had its documentation and translations stripped out.  The reinstallation
would be skipped if for some reason it was no longer part of the fedora
base image.

This was misinterpreted when creating the Kickstart equivalent of the
fedora-toolbox OCI images [3] and that's how it ended up in the KIWI
description.

In reality, p11-kit gets pulled in as a dependency of ca-certificates,
gnutls, etc. and that's sufficient.

Fallout from d27e4c3f6c3e13bcefe9d1d904386794140d8a91

[1] https://pagure.io/fedora-comps/

[2] https://src.fedoraproject.org/container/fedora-toolbox

[3] fedora-kickstarts commit 48e2c3b5598de32f
    https://pagure.io/fedora-kickstarts/c/48e2c3b5598de32f
    https://pagure.io/fedora-kickstarts/pull-request/964
    https://pagure.io/releng/issue/11399

https://pagure.io/fedora-kiwi-descriptions/pull-request/189