LiveInstall: drop 'anaconda' package
The package called 'anaconda' is empty and only depends on two
things, anaconda-gui and anaconda-tui. Since anaconda 43.32,
anaconda runs the GTK UI if it's installed; it will only run the
web UI if anaconda-gui is not installed. So by including anaconda
we are forcing lives to run GTK UI and not really gaining
anything - we don't want anaconda-tui on lives anyway, I don't
think.

The anaconda-live package will pull in anaconda-webui, so all we
need to do is stop including anaconda. This will cause all live
images to use web UI, which was the intent of
https://github.com/rhinstaller/anaconda/pull/6530 . I guess if
any lives really want/need to use GTK UI, they can put
anaconda-gui into their profiles.

Signed-off-by: Adam Williamson <awilliam@redhat.com>