eln/wsl: drop the SELinux packages from the ELN WSL image
WSL does not support SELinux, so at best it's pointless to install. This
also actively breaks WSL 1 users, although I don't think we should
suggest that we support that, and if the policy on the builder doesn't
match the target things break during the build.

It also reduced the build time from ~80 seconds to ~70 seconds for me,
so that's nice.

See: 39b379f7d ("wsl: drop the SELinux packages from the WSL base image")
Signed-off-by: Jeremy Cline <jeremycline@linux.microsoft.com>