teams/cloud: Drop /boot partition and use a Btrfs subvolume for /boot
This lets us free up the space allocated for the partition and gives
us dynamically allocated space for boot data like we have with the
rest of the operating system.

We do not do this for s390x because zipl does not support it.

Reference: https://fedoraproject.org/wiki/Changes/BtrfsBootForCloud