teams/lxde: Do not request nonexistent fedora-release-lxde
The Fedora LXDE spin does not have a fedora-release variant,
so the image build fails when trying to request one.

This is an error in fedora-release and will (hopefully) be eventually
fixed. For now, disable the block requesting it so we can restore it
later.

(cherry picked from commit b6d7ae177da96ac9d472a1309b96ae9514ba749f)